requirements.txt
openai-agents
fastapi
uvicorn
python-multipart
pypdf
pydantic
python-dotenv